Fix rist_sender_send_data_balanced in the case when no peer was selected
When running rist in load balanced mode rist_sender_send_data_balanced is occasionally not selecting any peer, thereby resulting in the packet not getting send (and no counters getting incremented, making recovery impossible).
This fixes that by recursing back into the function when this happens.
We keep a peercount of peers we could send data to in order to prevent an endless loop