Fix loadbalanced listening mode.
In listening mode currently we count each connection as a peer, so if
multiple clients connect traffic gets loadbalanced accross all those clients.
In stead of each client getting the full stream clients will receive a
load balanced partial stream.

We try to address that by keeping a linked list of child clients in the
listening peer, that we loop over when sending data