defined(__linux) may not imply linux/if_alg.h
When I try to compile against musl libc on Ubuntu 20.04, linux/if_alg.h
is not available.

This change simply uses the fallback code if the header isn't found at
configure-time.

I'm not sure if there's some missing package that would provide this.  I
searched but did not spot a candidate.